Refactor old ProjectStepTransition -> new ProjectWorkflowTransition model/resolver
